Exit read loop if context cancels
If the context has been canceled, the read loop goroutine should exit.
Because the read loop blocks on reading from the network connection,
there could be an arbitrary delay between the time someone calls
cancelFn() and the time this goroutine actually exits.  At some point
it'd be nice to use conn.SetReadDeadline() to make sure that we only
block on reads for a limited amount of time.